requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
